+++ test/mtx/mtx-remote.test.js
@@ -3,7 +3,8 @@
 const cds = require("@sap/cds");
 const supertest = require("supertest");
 
-global.fetch = jest.fn();
+const fetch = require("node-fetch");
+jest.mock("node-fetch", () => jest.fn());
 
 const util = require("../_env/util/request");